requirements.txt
beautifulsoup4==4.14.2
selenium==4.38.0
requests==2.32.5
pymilvus==2.6.3
marshmallow==3.26.1
environs==14.5.0
sentence-transformers==5.1.2
ollama==0.6.0
gradio==5.49.1
fastapi==0.121.0
uvicorn==0.38.0
pydantic==2.11.10
PyPDF2==3.0.1
langchain==1.0.3
langchain-core==1.0.3
langchain-community==0.4.1
pytesseract==0.0.13
Pillow==11.3.0